change HCAL Producer into class
